+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Added a graphical controller popout that can be accessed by clicking on the kmbmpdc icon in the menubar.
 - List of upcoming tracks is visible in the controller and the visbility of that list can be toggled on and off.
+- MPD database can be searched through the search UI activated by the magnifying glass button.
 
 ### Changed
 - The three-button menubar controls have been replaced with a single menubar icon indicating current playback status.